Update the validateBlock header to be exposed from BFT class
